backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.24.0
supabase>=2.0.0
earthengine-api==0.1.382
google-genai
httpx>=0.25.0
pydantic==2.5.0
python-dotenv==1.0.0
requests>=2.28.0asyncpg>=0.29.0
